requirements.txt
# Core ML and Computer Vision Dependencies
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
timm>=0.9.0

# Image Processing
Pillow>=9.0.0

# Visualization
matplotlib>=3.5.0

# Web Framework Dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
flask>=2.3.0
python-multipart>=0.0.6

# Utilities
numpy>=1.21.0
requests>=2.28.0

# Optional: For better performance and additional features
# opencv-python>=4.7.0  # Alternative image processing library
# scipy>=1.9.0          # Scientific computing utilities

# Note: Models will be cached locally in ./models/ directory
# Run download_models.py once to cache all model files locally
